requirements.txt
discord.py
pyyaml
yt-dlp
openai